cli: Call EZSP.close() on cli command completion
When testing "bellows form" locally, I noticed that sometimes the
command would not exit on it's own. When investigating, I noticed that
the utility shutdown() function incorrectly referred to the function app
rather than the local app created in async_inner(). The resultant
AttributeError was previously unnoticed because shutdown() silently
catches all exceptions.

This patch adds a reference to application as a local variable in the
top-level utility function, allowing close() to be called correctly.